+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v1.3.5 - Unreleased
 ----
  * Fix type verification in resolver arguments, GraphQL Schema must do this
  * Fix bug with redis pub/sub handler and socket timeouts
+ * Fix bug when use search in node pagination with  chars like `'%_`
 
 v1.3.4 - 2019-05-03
 ----